created LocalCommunicator out of DummyComm, linking dummycomm to localcomm to not break things
